#4332 moved specs also
